requirements.txt
# Call Insights Engine - Dependencies
# Python 3.10+ required

# Google Cloud & AI
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
google-auth>=2.23.0
vertexai>=1.38.0

# Google Agent Development Kit (ADK)
# Install from: pip install google-adk
# Or: pip install git+https://github.com/google/adk-python.git

# HTTP & Networking
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Audio Processing (optional - for local duration detection)
# pydub>=0.25.1
# ffmpeg-python>=0.2.0

# Environment & Config
python-dotenv>=1.0.0

# CLI & Utilities
tqdm>=4.66.0
click>=8.1.0

# Logging & Monitoring
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking
mypy>=1.7.0
types-requests>=2.31.0
